extract ScanIndex()
